+++ src/test/unit/index.ts
@@ -1,2 +1,5 @@
 import './exampleUnitTest.test.js';
 import './otherExtensions.test.js';
+import './panelModel.test.js';
+import './panelParser.test.js';
+import './converter.test.js';